2017.01.20

+++ app/Http/Controllers/SfusiRemoveController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 
 use App\sfusiStock;
+use App\shipStock;
 use DB;
 
 use Session;
@@ -95,13 +96,94 @@ class SfusiRemoveController extends Controller {
 		//dd($cb_to_remove_array_unique);
 		if (isset($cb_to_remove_array)) {
 			foreach ($cb_to_remove_array as $line) {
-					// dd($line['cartonbox']);
-					// if ($line == 'cartonbox') {
-						// dd($line);
-						$results = sfusiStock::where('cartonbox', '=', $line['cartonbox'])->delete();
-					// }
+				
+				$results = sfusiStock::where('cartonbox', '=', $line['cartonbox'])->delete();
 				
 			}
+			Session::set('cb_to_remove_array', null);
+			$msg = "All scanned Cartonbox succesfuly removed from Stock";
+			return view('Remove.success',compact('msg'));
+		}
+
+		Session::set('cb_to_remove_array', null);
+		$msg = "List of Cartonbox to delete is empty";
+		return view('Remove.success',compact('msg'));		
+	}
+
+	public function destroycb2(Request $request)
+	{
+
+		$cb_to_remove_array = Session::get('cb_to_remove_array');
+		// dd($cb_to_remove_array);
+
+		if ($cb_to_remove_array != NULL) {
+
+			$cb_to_remove_array_unique = array_map("unserialize", array_unique(array_map("serialize", $cb_to_remove_array)));
+			Session::set('cb_to_remove_array', null);
+			Session::set('cb_to_remove_array', $cb_to_remove_array_unique);
+
+			$cb_to_remove_array = $cb_to_remove_array_unique;
+		}
+
+		$msg1 = "";
+
+		//dd($cb_to_remove_array_unique);
+		if (isset($cb_to_remove_array)) {
+			foreach ($cb_to_remove_array as $line) {
+				
+				// $ship_table = shipStock::where('cartonbox', '=', $line['id']);
+				$sfusi_table = DB::connection('sqlsrv')->select(DB::raw("SELECT * FROM sfusiStock WHERE cartonbox = ". $line['cartonbox']));
+
+				$style = $sfusi_table[0]->style;
+				$size = $sfusi_table[0]->size;
+				$color = $sfusi_table[0]->color;
+				// var_dump($style);
+
+				$ship_table = DB::connection('sqlsrv')->select(DB::raw("SELECT cartonbox FROM shipStock WHERE style = '".$style."' AND color = '".$color."' AND size = '".$size."'"));
+				// dd($ship_table);
+
+				if (empty($ship_table)){
+
+					// Add to ship
+					try {
+						$table = new shipStock;
+
+						$table->cartonbox = $sfusi_table[0]->cartonbox;
+						$table->po = $sfusi_table[0]->po;
+						$table->po_status = $sfusi_table[0]->po_status;
+
+						$table->style = $sfusi_table[0]->style;
+						$table->color = $sfusi_table[0]->color;
+						$table->colordesc = $sfusi_table[0]->colordesc;
+						$table->size = $sfusi_table[0]->size;
+
+						$table->qty = $sfusi_table[0]->qty;
+						$table->standard_qty = $sfusi_table[0]->standard_qty;
+
+						$table->location = "SHIP"; 
+
+						$table->save();
+					}
+					catch (\Illuminate\Database\QueryException $e) {
+						$msg = "Problem to save cb in shipstock table";
+						return view('Remove.error',compact('msg'));
+					}
+
+					// Delete from sfusiStock
+					$results = sfusiStock::where('cartonbox', '=', $line['cartonbox'])->delete();	
+					
+				} else {
+					$msg1 = $msg1." ".$line['cartonbox'];
+					
+				}
+			}
+
+			if ($msg1 != "") {
+				Session::set('cb_to_remove_array', null);
+				$msg = $msg1." This cartonbox have SKU that already exist in SHIP table";
+				return view('Remove.error',compact('msg'));
+			}
+			
 			Session::set('cb_to_remove_array', null);
 			$msg = "All scanned Cartonbox succesfuly removed from Stock";
 			return view('Remove.success',compact('msg'));
